// src/types/indexType.ts
export interface activeType {
  activeRoom: number;
  activeWarhouse: number;
  activeParking: number;
  activeBathroom: number;
  activeBathroomType: number;
  activeFilterElevator: number;
  activeFloor: number;
  activeCollingSystem: number;
  activeHeatingSystem: number;
  activeFloorMaterial: number;
}